split the DeContextMapKey in CE and EE classes
